Refactoring of the components directory
We now longer depend on the update_components.sh script. We have 3 components as git submodules, namely:
	1) esp-nn: https://github.com/espressif/esp-nn.git
	2) esp32-akri: https://github.com/nubificus/esp32-akri.git
	3) ota-service: https://github.com/nubificus/ota-service.git

There is also the tfmicro component which is the directory found in:
	* https://github.com/espressif/esp-tflite-micro/tree/4b9a1732e2f0901ec33ab8337aad5469691d4e98/components/tflite-lib

Supposedly, the script sync_from_tflite_micro.sh is used to occasionaly update the tflite-lib directory.
Therefore, I created a slighlty modified one in this project's scripts directory to do the same.
This should update the tfmicro component but the structure is not correct(i.e. missing libs, etc.).
For the time being ignore this script and use the existing tfmicro component, but the scipt is in:
	* https://github.com/espressif/esp-tflite-micro/blob/4b9a1732e2f0901ec33ab8337aad5469691d4e98/scripts/sync_from_tflite_micro.sh

To sum up, by using git clone --recursive you should be able to instantly do idf.py build.