Made esp32 the server and added environment variable for ota service
The esp32 is now a TCP server that can service multiple TCP clients.
A TCP client connect to the esp32, sends a request byte followed
by an input file and the esp32 receives it, does the inference
and finally sends back the results and inference time. The client
can continue sending data for inference to esp32.

The other change is that now OTA service is added only if the
STOCK environment variable is set.