Made the project almost independent of the model
The project can now work with other models as well by changing
only the micro_op_resolver in main_functions.cpp.

From now on the raw input data are copied on the input tensor and the output tensor
is handled by a general Dequantize function. We might need to add a similar Quantize
function for the input tensor but since it is uncommon to quantize the input,
we shall examine this in the future.